+++ src/Features/ERDDAP/Platform/Observations/Table/all.spec.tsx
@@ -9,10 +9,15 @@ import { ErddapAllObservationsTable } from "./all"
 
 describe("<ErddapAllObservationsTable>", () => {
   it("Should show all observations for appropriate platform", () => {
+    // Need to have a div in the body for the tooltip to attach to
+    const div = document.createElement("div")
+    document.body.appendChild(div)
+
     const enzymeWrapper = mount(
       <MemoryRouter>
         <ErddapAllObservationsTable platform={platform} unit_system={UnitSystem.english} />
-      </MemoryRouter>
+      </MemoryRouter>,
+      { attachTo: div }
     )
 
     expect(enzymeWrapper.find("a").length).toBe(16)
@@ -25,7 +30,7 @@ describe("<ErddapAllObservationsTable>", () => {
 const platform: PlatformFeatureWithDatasets = {
   geometry: {
     type: "Point",
-    coordinates: [-73.73, 40.88]
+    coordinates: [-73.73, 40.88],
   },
   id: "EXRX",
   type: "Feature",
@@ -40,15 +45,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "northward_wind",
           short_name: null,
           long_name: "Wind Speed, Meridional",
-          units: "m s-1"
+          units: "m s-1",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "wspv",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: -0.699999988079071,
@@ -58,15 +63,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "eastward_wind",
           short_name: null,
           long_name: "Wind Speed, Zonal",
-          units: "m s-1"
+          units: "m s-1",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "wspu",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -76,15 +81,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "surface_altitude",
           short_name: null,
           long_name: "Water Level",
-          units: "m"
+          units: "m",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "tide",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -94,15 +99,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "tendency_of_air_pressure",
           short_name: null,
           long_name: "Pressure Tendency",
-          units: "hPa"
+          units: "hPa",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "ptdy",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -112,15 +117,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "visibility_in_air",
           short_name: "VIS",
           long_name: "Visibility",
-          units: "meters"
+          units: "meters",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "vis",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -130,15 +135,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "dew_point_temperature",
           short_name: "DEWP",
           long_name: "Dewpoint Temperature",
-          units: "celsius"
+          units: "celsius",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "dewp",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: 4.699999809265137,
@@ -148,15 +153,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "sea_surface_temperature",
           short_name: null,
           long_name: "SST",
-          units: "degree_C"
+          units: "degree_C",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "wtmp",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -166,15 +171,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "air_temperature",
           short_name: "AT",
           long_name: "Air Temperature",
-          units: "celsius"
+          units: "celsius",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "atmp",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -184,15 +189,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "air_pressure_at_sea_level",
           short_name: null,
           long_name: "Air Pressure",
-          units: "hPa"
+          units: "hPa",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "bar",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -202,15 +207,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "sea_surface_wave_to_direction",
           short_name: null,
           long_name: "Wave Direction",
-          units: "degrees_true"
+          units: "degrees_true",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "mwd",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -220,15 +225,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "sea_surface_swell_wave_period",
           short_name: null,
           long_name: "Wave Period, Dominant",
-          units: "s"
+          units: "s",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "apd",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -238,15 +243,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "sea_surface_swell_wave_period",
           short_name: null,
           long_name: "Wave Period, Dominant",
-          units: "s"
+          units: "s",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "dpd",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: null,
@@ -256,15 +261,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "sea_surface_wave_significant_height",
           short_name: null,
           long_name: "Wave Height",
-          units: "m"
+          units: "m",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "wvht",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: 3.0,
@@ -274,15 +279,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "wind_speed_of_gust",
           short_name: "WGST",
           long_name: "Wind Gust",
-          units: "m/s"
+          units: "m/s",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "gst",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: 2.0,
@@ -292,15 +297,15 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "wind_speed",
           short_name: "WSPD",
           long_name: "Wind Speed",
-          units: "m/s"
+          units: "m/s",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "wspd",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
+        start_time: "1970-02-26T20:00:00Z",
       },
       {
         value: 20.0,
@@ -310,16 +315,16 @@ const platform: PlatformFeatureWithDatasets = {
           standard_name: "wind_from_direction",
           short_name: "WD",
           long_name: "Wind Direction",
-          units: "degrees"
+          units: "degrees",
         },
         server: "https://coastwatch.pfeg.noaa.gov/erddap",
         variable: "wd",
         constraints: {
-          "station=": "44022"
+          "station=": "44022",
         },
         dataset: "cwwcNDBCMet",
-        start_time: "1970-02-26T20:00:00Z"
-      }
+        start_time: "1970-02-26T20:00:00Z",
+      },
     ],
     attribution: [],
     mooring_site_desc: "Execution Rocks Long Island Sound",
@@ -328,6 +333,6 @@ const platform: PlatformFeatureWithDatasets = {
     uscg_light_letter: null,
     uscg_light_num: null,
     watch_circle_radius: null,
-    programs: []
-  }
+    programs: [],
+  },
 }