+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Changes:
 - Remove signal flags.
 - Allow multiple data types to be specified in current conditions table and show only the first one matched.
 - Show only one of a group of data types in the current (latest) conditions view.
+- Show only a 15 min range of time in the current conditions table.
+- Adds a tooltip popup for condition tables with the time of reading.
 
 Fixes: